feat: Extend palindromeCheck functionality to sentences
Returns longest unique palindrome in a sentence.
A sentence is defined as a single string of words separated by whitespace.
If there are 2 or more largest palindromes, they are all returned.
The getLongestPalindrome method also ignores non alphabetical characters.

Create a method which checks if a word is a palindrome

A palindrome is at minimum 3 characters long and is the same forwards as it is backwards
Spaces should not be ignored
The method should be case insensitive

For example: "Dad" is a palindrome, but "race car" is not.

Part 2:
Create a program that finds the longest unique Palindrome in a sentence
A sentence is defined as a single string of words separated by whitespace
If there are 2 or more largest palindromes, both should be returned
The program should ignore all non-alphabet characters